Remove the moved nexus.rb states
